front step-4 validation with unit testing
