+++ app/Http/Controllers/Front/AccountsController.php
@@ -20,6 +20,7 @@ use App\Shop\Customers\Repositories\CustomerRepository;
 use App\Shop\Orders\Repositories\Interfaces\OrderRepositoryInterface;
 use App\Shop\Couriers\Repositories\Interfaces\CourierRepositoryInterface;
 use App\Shop\Customers\Repositories\Interfaces\CustomerRepositoryInterface;
+use Illuminate\Support\Facades\Input;
 
 class AccountsController extends Controller
 {
@@ -289,22 +290,44 @@ class AccountsController extends Controller
     
     public function updateTeethImages(Request $request)
     {
-        if(count($request->file('images')) == 0)
-        {
-            return $this->sendResponse(false, "Teeth Images are required");
-        }
+        if(Input::get('submit')) {
+          $description = $request->description;
+            if($request->file('image') == 0)
+            {
+                return $this->sendResponse(false, "Teeth Images are required");
+            }
 
-        $user = $this->loggedUser();
-        foreach($request->images as $image)
-        {
-            CustomerImage::create([
-                'customer_id' => $user->id,
-                'image' => $this->saveImage($image, 'images')
-            ]);
-        }
-
-        event(new AddNotification($user->id, 1, 'You have uploaded teeth images.'));        
-        return $this->sendResponse(true, 'Image uplaoded');
+            $user = $this->loggedUser();
+           
+                CustomerImage::create([
+                    'customer_id' => $user->id,
+                    'image' => $this->saveImage($request->file('image'), 'images'),
+                    'description' => $description
+                ]);
+    
+            event(new AddNotification($user->id, 1, 'You have uploaded teeth images.'));        
+            return $this->sendResponse(true, 'Image uplaoded');
+      
+      }else {
+            $customers = CustomerImage::find($request->doc_id_name);
+            if(empty($customers ) && $customers ->count() == 0){
+                return $this->sendResponse(false, "Teeth Images are required");
+            }
+            //Merge new data from request
+            $updateArray = $request->all();
+            if(!empty($updateArray['image'])){
+              $path = $this->saveImage($request->file('image'),'images');
+              $updateArray['image'] = $path;
+            }
+            $customers->fill($updateArray);
+           
+            if($customers->save()){
+                event(new AddNotification($customers, 1, 'You have uploaded teeth images.'));        
+                return $this->sendResponse(true, 'Image uplaoded');
+            } else{
+                return $this->sendResponse(false, "Teeth Images are required");
+            }
+        } 
     }
 
     public function removeTeethImage($image)
@@ -314,7 +337,6 @@ class AccountsController extends Controller
         $teethImage->delete();
 
         return $this->sendResponse(true, 'Image removed');
-
     }
 
     public function updatePassword(Request $request)
